jfn_cef: split client and client_impl into submodules
